requirements.txt
numpy
Pillow
open3d==0.19.0
